Extract ImageFlipper to its own file
